requirements.txt
alembic==1.8.1
blinker==1.5
certifi==2022.6.15
click==8.1.3
dnspython==2.2.1
dominate==2.7.0
elastic-transport==8.1.2
elasticsearch==8.3.3
email-validator==1.2.1
Flask==2.1.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
SQLAlchemy==1.4.39
urllib3==1.26.11
visitor==0.1.3
Werkzeug==2.2.0
WTForms==3.0.1
zipp==3.8.1
